src/test/java/_0851_0900/_852_Peak_Index_in_a_Mountain_Array/SolutionITest.java
package _0851_0900._852_Peak_Index_in_a_Mountain_Array;

import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.assertEquals;

/**
 * @author jacka
 * @version 1.0 on 9/6/2020
 */
class SolutionITest {

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase1(Solution solution) {
    final int arr[] = {0, 1, 0}, expected = 1, actual = solution.peakIndexInMountainArray(arr);
    assertEquals(expected, actual);
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase2(Solution solution) {
    final int arr[] = {0, 2, 1, 0}, expected = 1, actual = solution.peakIndexInMountainArray(arr);
    assertEquals(expected, actual);
  }

  @ParameterizedTest
  @MethodSource("solutionStream")
  void testOnlineCase3(Solution solution) {
    final int arr[] = {0, 10, 5, 2}, expected = 1, actual = solution.peakIndexInMountainArray(arr);
    assertEquals(expected, actual);
  }

  static Stream<Solution> solutionStream() {
    return Stream.of(
        new SolutionI()
    );
  }
}